requirements.txt
# ============================================================
# NewPick Video Pipeline — Python Dependencies
# Install: uv pip install -r requirements.txt
# ============================================================

# ---------- Google AI / Gemini ----------
google-genai>=0.7.0
google-generativeai>=0.8.0
google-cloud-speech>=2.26.0
google-api-python-client>=2.100.0
google-auth>=2.30.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0

# ---------- OpenRouter / HTTP ----------
httpx>=0.27.0
openai>=1.0.0

# ---------- Image Processing ----------
pillow>=10.0.0
opencv-python-headless>=4.8.0
numpy>=1.24.0

# ---------- Video / Audio ----------
moviepy>=1.0.3
imageio>=2.31.0
imageio-ffmpeg>=0.4.9
pydub>=0.25.1

# ---------- Subtitles / Transcription ----------
pysrt>=1.1.2

# ---------- Image Crawling ----------
icrawler>=0.6.6
beautifulsoup4>=4.12.0
lxml>=4.9.0

# ---------- Web / Networking ----------
requests>=2.31.0
selenium>=4.15.0
urllib3>=2.0.0

# ---------- Config / Env ----------
python-dotenv>=1.0.0
pyyaml>=6.0.0
pydantic>=2.0.0

# ---------- Telegram Bot ----------
python-telegram-bot[job-queue]>=20.0

# ---------- Fonts ----------
uharfbuzz>=0.39.0

# ---------- Utilities ----------
tqdm>=4.66.0
tenacity>=8.2.0
Flask>=3.0.0
PyJWT>=2.8.0
anyio>=4.0.0
